refactor: 현재 월에 대한 달력을 그리기 위한 로직을 훅으로 분리
