+++ features/main/calendar/molecules/calendar.tsx
@@ -1,34 +1,13 @@
 'use client';
 
-import dayjs from 'dayjs';
-import { useAtomValue } from 'jotai';
-
-import { calendarDateAtom } from '@/store';
+import { useCalendarRendaringData } from '@/hooks/use-calendar-rendering-data';
 import { css } from '@/styled-system/css';
 
 import { CalendarItem, DayLabels } from '../atoms';
 import { CalendarHeader } from './calendar-header';
 
-const CALENDAR_FORMAT_5_COLUMNS = 35;
-const CALENDAR_FORMAT_6_COLUMNS = 42;
-
 export const Calendar = () => {
-  const currentDate = useAtomValue(calendarDateAtom);
-
-  const { year, month } = currentDate;
-  const dateObj = dayjs(`${year}-${month}-01`);
-  const startPoint = dateObj.get('day');
-  const totalDays = dateObj.add(1, 'month').subtract(1, 'day').get('date');
-  const endPoint = startPoint + totalDays - 1;
-  const squares = Array.from(
-    {
-      length:
-        endPoint >= CALENDAR_FORMAT_5_COLUMNS
-          ? CALENDAR_FORMAT_6_COLUMNS
-          : CALENDAR_FORMAT_5_COLUMNS,
-    },
-    (_, i) => i + 1,
-  );
+  const [squares, startPoint, endPoint] = useCalendarRendaringData();
 
   return (
     <div className={calendarContainer}>
